requirements.txt
asgiref==3.9.1
Django==5.2.5
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==37.5.3
psycopg2-binary==2.9.10
PyJWT==2.10.1
ruff==0.12.9
sqlparse==0.5.3
tzdata==2025.2
